SPM12 Results Report

SPM12 Results Report

spm12_results(
  ...,
  add_spm_dir = TRUE,
  spmdir = spm_dir(verbose = verbose, install_dir = install_dir),
  clean = TRUE,
  verbose = TRUE,
  display = FALSE,
  desktop = FALSE,
  install_dir = NULL
)

build_spm12_results(
  spm,
  units = c("Volumetric", "Scalp-Time", "Scalp-Frequency", "Time-Frequency",
    "Frequency-Frequency"),
  result_format = c("none", "ps", "eps", "png", "pdf", "jpg", "tif", "fig", "csv",
    "nidm"),
  write_images = c("none", "threshold_spm", "binary_clusters", "nary_clusters"),
  contrast_list = NULL,
  image_basename = NULL,
  ...
)

Arguments

...

Arguments passed to matlabbatch_to_script

add_spm_dir

Add SPM12 directory from this package

spmdir

SPM dir to add, will use package default directory

clean

Remove scripts from temporary directory after running

verbose

Print diagnostic messages

display

Run run_matlab_script with the display option on, which is required in some cases of output. May fail if no displays are available.

desktop

Run run_matlab_script with the desktop option on. May fail if no displays are available.

install_dir

directory to download SPM12

spm

Path to SPM.mat file

units

Units of the data

result_format

Output format to save, if any

write_images

Type of images to write out

contrast_list

List of contrasts (or just one), to pass to spm12_contrast_query_list

image_basename

Base stub of filenames, if any are to be written out

Value

A list of output and results