SPM12 Slice Timing Correction

Performs SPM12 slice timing correction on images

spm12_slice_timing.deprecated(filename, time_points = NULL, nslices, tr,
  ta = tr - tr/nslices, slice_order = 1:nslices, ref_slice, prefix = "a",
  add_spm_dir = TRUE, spmdir = spm_dir(), clean = TRUE, verbose = TRUE,
  outdir = tempdir(), retimg = FALSE, reorient = FALSE, ...)

Arguments

filename
Files to be slice-time corrected

time_points
Time points to run slice-time correction. If filename is a 4D file, then will do all the time points. Otherwise, filename must be a character vector of 3D files or a list of 3D nifti objects.

nslices
Number of slices in the images

tr
Repetition time (in seconds)

ta
Time between the first and the last slice within one scan

slice_order
Order slices were taken (if not specified, assumed ascending), bottom slice = 1

ref_slice
Reference slice

prefix
Prefix to append to front of image filename

add_spm_dir
Add SPM12 directory from this package

spmdir
SPM dir to add, will use package default directory

clean
Remove scripts from temporary directory after running

verbose
Print diagnostic messages

outdir
Directory to copy results

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readNIfTI.

...
Arguments passed to run_spm12_script

Value

Output filenames