SPM12 Smooth

Performs SPM12 Smoothing on an Image

spm12_smooth(
  filename,
  retimg = FALSE,
  fwhm = 8,
  dtype = c("SAME", "UINT8", "INT16", "INT32", "FLOAT32", "FLOAT64"),
  implicit_mask = FALSE,
  prefix = "s",
  add_spm_dir = TRUE,
  spmdir = spm_dir(verbose = verbose, install_dir = install_dir),
  clean = TRUE,
  verbose = TRUE,
  reorient = FALSE,
  install_dir = NULL,
  ...
)

Arguments

filename

File to be smoothed

retimg

Logical indicating if image should be returned or result from run_matlab_script

fwhm

Full-Width Half Max to smooth

dtype

data type for the output format

implicit_mask

Should an implicit mask be used. An "implicit mask" is a mask implied by a particular voxel value (0 for images with integer type, NaN for float images).

prefix

Prefix to append to front of image filename

add_spm_dir

Add SPM12 directory from this package

spmdir

SPM dir to add, will use package default directory

clean

Remove scripts from temporary directory after running

verbose

Print diagnostic messages

reorient

if retimg=TRUE pass to readNIfTI

install_dir

directory to download SPM12

...

Arguments passed to run_spm12_script readNIfTI

Value

Result from run_matlab_script or nifti file, depending on retimg