Build Conditions for SPM12 first level model

Build Conditions for SPM12 first level model

spm12_condition(name, onset, duration, time_mod_order = 0,
  param_mod = NULL, orth = TRUE)

spm12_condition_list(cond)

Arguments

name
Name of the condition

onset
vector of onset of the condition

duration
vector of duration of the condition, must be the same length as onset

time_mod_order
time modulation order. This option allows for the characterization of linear or nonlinear time effects. Zero means no modulation

param_mod
parametric modulation. Not currently supported in spm12r

orth
Orthogonalize the regressors within trial types.

cond
List of conditions

Value

A list of objects, each with a name, onset, duration, and other condition values.

Examples

res = spm12_condition(
name = "condition1", 
onset = c(0, 2, 4, 6, 8),
duration = rep(1, 5) )
print(res)
#> $name
#> [1] "'condition1'"
#> 
#> $onset
#> [1] "[0, 2, 4, 6, 8];"
#> 
#> $duration
#> [1] "[1, 1, 1, 1, 1];"
#> 
#> $tmod
#> [1] 0
#> 
#> $pmod
#> [1] "struct('name', {}, 'param', {}, 'poly', {})"
#> 
#> $orth
#> [1] 1
#> 
L = list(
cond1 = list(onset = c(0, 2, 4, 6, 8), duration = rep(1, 5)), 
cond2 = list(onset = c(0, 2, 4, 6, 8) + 1, duration = rep(1, 5))
)
res = spm12_condition_list(L)
print(res) 
#> $`(1)`
#> $`(1)`$name
#> [1] "'cond1'"
#> 
#> $`(1)`$onset
#> [1] "[0, 2, 4, 6, 8];"
#> 
#> $`(1)`$duration
#> [1] "[1, 1, 1, 1, 1];"
#> 
#> $`(1)`$tmod
#> [1] 0
#> 
#> $`(1)`$pmod
#> [1] "struct('name', {}, 'param', {}, 'poly', {})"
#> 
#> $`(1)`$orth
#> [1] 1
#> 
#> 
#> $`(2)`
#> $`(2)`$name
#> [1] "'cond2'"
#> 
#> $`(2)`$onset
#> [1] "[1, 3, 5, 7, 9];"
#> 
#> $`(2)`$duration
#> [1] "[1, 1, 1, 1, 1];"
#> 
#> $`(2)`$tmod
#> [1] 0
#> 
#> $`(2)`$pmod
#> [1] "struct('name', {}, 'param', {}, 'poly', {})"
#> 
#> $`(2)`$orth
#> [1] 1
#> 
#>