SPM12 FMRI Estimation

SPM12 FMRI Estimation

spm12_fmri_est(..., add_spm_dir = TRUE, spmdir = spm_dir(verbose = verbose),
  clean = TRUE, verbose = TRUE)

build_spm12_fmri_est(spm, write_residuals = FALSE, method = c("Classical",
  "Bayesian", "Bayesian2"), bayesian = list(space.volume.block_type =
  "Slices", signal = "UGL", ARP = 3, noise.UGL = 1, LogEv = "No", anova.first =
  "No", anova.second = "Yes", gcon = list(name = "", convec = numeric(0))), ...)

Arguments

...
Arguments passed to matlabbatch_to_script

add_spm_dir
Add SPM12 directory from this package

spmdir
SPM dir to add, will use package default directory

clean
Remove scripts from temporary directory after running

verbose
Print diagnostic messages

spm
Path to SPM.mat file

write_residuals
Should residuals be written?

method
Method for model estimation

bayesian
If method = "Bayesian", this is for a 1st level model Bayesian estimation and this list specifies the parameters

Value

A list of output and results