SPM BWLabel Clusters of Certain Size

Get Cluster of certain size from spm_bwlabel

spm_bwlabel(
  infile,
  outfile = NULL,
  retimg = TRUE,
  k = NULL,
  topN = NULL,
  margin = NULL,
  binary = TRUE,
  spmdir = spm_dir(verbose = verbose, install_dir = install_dir),
  reorient = FALSE,
  verbose = TRUE,
  install_dir = NULL
)

Arguments

infile
input filename

outfile
output filename

retimg
Return the image instead of matlab output

k
Minimum cluster size needed

topN
Top number of clusters kept (used if k is NULL)

margin
Margin to loop over if wanted in 2D

binary
(logical) Should the result be binary or numbered with cluster.

spmdir
SPM directory (for MATLAB)

reorient
If retimg, then this argument is passed to readNIfTI

verbose
Print Diagnostics

install_dir
directory to download SPM12

Value

Output from run_matlab_script or nifti object, depending on retimg

Result from run_matlab_script

Note

Taken from http://en.wikibooks.org/wiki/SPM/How-to#How_to_remove_clusters_under_a_certain_size_in_a_binary_mask.3F

Examples

if (matlabr::have_matlab()) { 
d = rep(25, 3)
arr = array(rbinom(prod(d), size = 1, prob = 0.5), dim = d)
nim = oro.nifti::nifti(arr)
res = spm_bwlabel(nim, k = 1, binary = FALSE)
tab = table(c(res))
tab
out_arr = as(res, "array")
}

#> cloning into '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/Rtmpjs2B3J/temp_libpath115154e4dda52/spm12r/spm12'...
#> Receiving objects:   1% (66/6503),   17 kb
#> Receiving objects:  11% (716/6503), 2956 kb
#> Receiving objects:  21% (1366/6503), 3292 kb
#> Receiving objects:  31% (2016/6503), 9322 kb
#> Receiving objects:  41% (2667/6503), 10841 kb
#> Receiving objects:  51% (3317/6503), 16950 kb
#> Receiving objects:  61% (3967/6503), 35762 kb
#> Receiving objects:  71% (4618/6503), 37953 kb
#> Receiving objects:  81% (5268/6503), 69757 kb
#> Receiving objects:  91% (5918/6503), 71853 kb
#> Receiving objects: 100% (6503/6503), 129904 kb, done.

#> # K = 1

#> # Script is located at /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpL8mrxo/file1154640564d74.m

#> Command run is:

#> matlab  -nodesktop -nosplash   -nodisplay -r  "try, run('/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpL8mrxo/file1154640564d74.m'); catch err, disp(err.message); exit(1); end; exit(0);"

#> # Reading output file /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpL8mrxo/file11546309bbc0a.nii

#>